Create resources utils
